Here is the refactored code:
The refactored code is nearly identical to the original function, with small changes that address readability, maintainability, and structure. 

First, I renamed the function `relativeTime` to  `getRelativeTimeDiff` to make its purpose more explicit. I also defined an object `ms` which contains the number of milliseconds for each unit of time that the function uses in its calculations. 

Next, to avoid repetition, I created a helper function `getUnit` that calculates the value passed to it in the current time unit. This helper function is called by each of the `if` statements.

I also updated the pluralisation logic by removing the `isPlural` function and replaced with ternary conditions to check whether the value of a variable is '1'.

Lastly, I swapped the order of the equal sign and the values in the `else` block to provide accurate pluralization for the "year" result.

The refactored code should have a more modular design, it is more readable, performing functions that are critical to the operation of the code. The refactored code retains all of the functionality of the initial implementation, so it should perform equivalently and pass all tests.